refactor(gilgen): move GILGenLValue structure in other file
