+++ lib/GILGen/GILGenStmt.cpp
@@ -1,9 +1,9 @@
-#include "Context.hpp"
-#include "Scope.hpp"
-
 #include "ASTVisitor.hpp"
+#include "Context.hpp"
 #include "GILGen.hpp"
 #include "GILGenExpr.hpp"
+#include "GILGenLValue.hpp"
+#include "Scope.hpp"
 
 #include <llvm/ADT/SmallVector.h>
